+++ app/Http/Controllers/CommentController.php
@@ -19,14 +19,17 @@ class CommentController extends Controller
         ]);
         $comment = new Comment();
         $comment->article_id = $request->get('article');
-        $comment->user_id = \Auth::user()->id;
+        $comment->user_id = \Auth::id();
         $comment->content = $request->get('comment');
         $comment->save();
         return redirect('/article/'.$request->get('article'));
     }
 
     public function destroy($id){
-        Comment::find($id)->delete();
-        return redirect()->back();
+        if(\Auth::id() ==Comment::find($id)->user->id||\Auth::user()->hasRole('owner')||\Auth::user()->hasRole('admin')){
+            Comment::find($id)->delete();
+            return redirect()->back();
+        }
+
     }
 }
